finish tasks in readme file
