+++ sessions/backend-update-and-delete/products/components/Product/index.js
@@ -3,12 +3,17 @@ import { useRouter } from "next/router";
 import { StyledButton } from "../Button/Button.styled";
 import { ProductCard } from "./Product.styled";
 import Comments from "../Comments";
+import { useState } from "react";
+import ProductForm from "../ProductForm/index";
 
-export default function Product() {
+export default function Product({ onSubmit, onDelete }) {
+  const [isEditMode, setIsEditMode] = useState(false);
   const router = useRouter();
   const { id } = router.query;
 
   const { data } = useSWR(id ? `/api/products/${id}` : null);
+
+  if (!data) return;
 
   if (!data) {
     return <h1>Loading...</h1>;
@@ -25,6 +30,18 @@ export default function Product() {
       <StyledButton type="button" onClick={() => router.push("/")}>
         Back to all
       </StyledButton>
+      {isEditMode && (
+        <ProductForm onSubmit={onSubmit} heading="Update this Fish" />
+      )}
+      <button
+        type="button"
+        onClick={() => {
+          setIsEditMode(!isEditMode);
+        }}
+      >
+        Edit Product
+      </button>
+      <button onClick={() => onDelete(id)}>Delete</button>
     </ProductCard>
   );
 }
