+++ sessions/backend-update-and-delete/products/pages/[id].js
@@ -1,12 +1,9 @@
 import Product from "../components/product";
 import useSWRMutation from "swr/mutation";
-import { useState } from "react";
+
 import { useRouter } from "next/router";
-import ProductForm from "../components/ProductForm/index";
 
 export default function ProductDetailsPage() {
-  const [isEditMode, setIsEditMode] = useState(false);
-
   const router = useRouter();
   const {
     query: { id },
@@ -34,22 +31,16 @@ export default function ProductDetailsPage() {
     updateProduct
   );
 
-  async function handleEditProduct(productData) {
+  async function handleEditProduct(event) {
+    const formData = new FormData(event.target);
+    const productData = Object.fromEntries(formData);
     await trigger(productData);
     push("/");
   }
 
-  if (isMutating) {
-    return <h1>Submitting your changes.</h1>;
-  }
-
-  async function handleDeleteProduct() {
+  async function handleDeleteProduct(id) {
     const response = await fetch(`/api/products/${id}`, {
       method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(id),
     });
 
     if (response.ok) {
@@ -60,23 +51,12 @@ export default function ProductDetailsPage() {
     }
   }
 
+  if (isMutating) {
+    return <h1>Submitting your changes.</h1>;
+  }
   return (
     <>
-      {isEditMode && (
-        <ProductForm onSubmit={handleEditProduct} heading="Update this Fish" />
-      )}
-      <button
-        type="button"
-        onClick={() => {
-          setIsEditMode(!isEditMode);
-        }}
-      >
-        Edit Product
-      </button>
-      <button type="button" onClick={handleDeleteProduct}>
-        Delete Product
-      </button>
-      <Product />
+      <Product onSubmit={handleEditProduct} onDelete={handleDeleteProduct} />
     </>
   );
 }
